2020 Census Data for All U.S. Counties

A dataset containing demographic information by county for all counties in the United States from the 2020 Decennial Census. This dataset includes population counts by race and ethnicity for each of the 3,143 counties and county-equivalents in the United States, including Washington D.C. and territories.

census

Format

A data frame with 8 variables and counties as rows:

GEOID

5-digit FIPS code uniquely identifying each county (character)

NAME

County name and state (e.g., "Los Angeles County, California") (character)

total_pop

Total population count (integer)

white

Population identifying as White alone (integer)

black

Population identifying as Black or African American alone (integer)

native_american

Population identifying as American Indian and Alaska Native alone (integer)

asian

Population identifying as Asian alone (integer)

pacific_islander

Population identifying as Native Hawaiian and Other Pacific Islander alone (integer)

hispanic_latino

Population identifying as Hispanic or Latino of any race (integer)

Source

U.S. Census Bureau, 2020 Decennial Census, Summary File 1 Retrieved via the tidycensus R package from the U.S. Census Bureau API.

Details

The data comes from the 2020 Decennial Census Summary File 1 (SF1) tables:

P1_001N: Total population

P1_003N: White alone

P1_004N: Black or African American alone

P1_005N: American Indian and Alaska Native alone

P1_006N: Asian alone

P1_007N: Native Hawaiian and Other Pacific Islander alone

P2_002N: Hispanic or Latino (of any race)

Note that Hispanic/Latino is an ethnicity category that can apply to individuals of any race, so these counts may overlap with the race-specific categories.

The dataset includes all counties in the 50 states, Washington D.C., and U.S. territories (Puerto Rico, U.S. Virgin Islands, American Samoa, Guam, and Northern Mariana Islands).

Examples

# Load the data
data(census)

# View structure
str(census)
#> tibble [3,221 × 8] (S3: tbl_df/tbl/data.frame)
#>  $ name            : chr [1:3221] "Autauga County, Alabama" "Baldwin County, Alabama" "Barbour County, Alabama" "Bibb County, Alabama" ...
#>  $ total_pop       : int [1:3221] 58805 231767 25223 22293 59134 10357 19051 116441 13080 24275 ...
#>  $ white           : int [1:3221] 42160 189399 11317 16555 50663 2320 9752 80586 12633 22671 ...
#>  $ black           : int [1:3221] 11445 18217 11933 4413 845 7396 8430 25559 36 470 ...
#>  $ native_american : int [1:3221] 217 1582 116 60 337 72 29 487 18 40 ...
#>  $ asian           : int [1:3221] 881 2067 117 32 178 9 143 1190 9 60 ...
#>  $ pacific_islander: int [1:3221] 35 143 1 9 24 10 5 119 4 2 ...
#>  $ hispanic_latino : int [1:3221] 2117 12686 1510 740 5771 503 260 5010 83 475 ...

# Summary statistics
summary(census$total_pop)
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#>       64    10976    25828   103923    67260 10014009 

# Counties with highest population
census[order(-census$total_pop), ][1:10, c("NAME", "total_pop")]
#> Error in census[order(-census$total_pop), ][1:10, c("NAME", "total_pop")]: Can't subset columns that don't exist.
#> ✖ Column `NAME` doesn't exist.

# Calculate percentage demographics
census$pct_white <- round(census$white / census$total_pop * 100, 1)
census$pct_black <- round(census$black / census$total_pop * 100, 1)
census$pct_hispanic <- round(census$hispanic_latino / census$total_pop * 100, 1)

# Find counties with majority Hispanic population
hispanic_majority <- census[census$pct_hispanic > 50, c("NAME", "pct_hispanic")]
#> Error in census[census$pct_hispanic > 50, c("NAME", "pct_hispanic")]: Can't subset columns that don't exist.
#> ✖ Column `NAME` doesn't exist.

# Filter to specific state (using GEOID - first 2 digits are state FIPS)
california_counties <- census[substr(census$GEOID, 1, 2) == "06", ]
#> Warning: Unknown or uninitialised column: `GEOID`.
#> Error in census[substr(census$GEOID, 1, 2) == "06", ]: Can't subset rows with `substr(census$GEOID, 1, 2) == "06"`.
#> ✖ Logical subscript `substr(census$GEOID, 1, 2) == "06"` must be size 1 or 3221, not 0.

# Total US population
sum(census$total_pop)
#> [1] 334735155