Data from the Cooperative Election Study

Data for the approval ratings of voters to various government positions combined with the demographic background of the voter. See Kuriwaki (2024) for background and details. The raw code that was used to produce the data is accessible at Kuriwaki's Github.

ces

Format

A tibble with 701,955 observations and 17 variables:

case_id

integer variable for case number

year

integer variable for year of survey

state

character variable for state of residence for observation

sex

factor variable of "Female" and "Male"

age

integer variable for age in years

race

factor variable for racial identity

marstat

factor variable for marriage status

religion

factor variable for religious belief

faminc

ordered factor variable for yearly family income - takes values of "Less than 10k" "10k - 20k", "20k - 30k", "30k - 40k", "40k - 50k", "50k - 60k", "60k - 70k", "70k - 80k", "80k - 100k", "100k - 120k", "120k - 150k", and "150k+"

ideology

factor variable for self-reported ideology - takes values of "Very Liberal", "Liberal", "Moderate", "Conservative", "Very Conservative", and "Not Sure"

education

factor variable for level of education - takes values of "No HS", "High School Graduate", "Some College", "2-Year", "4-Year", and "Post-Grad"

news

factor variable for level news/current events interest

econ

factor variable for retrospective report on the past year's economy

approval

ordered factor variable of approval of president - takes values of "Strongly Approve", "Approve / Somewhat Approve", "Neither Approve Nor Disapprove", "Disapprove / Somewhat Disapprove", "Strongly Disapprove"

military

factor variable indicating whether the respondent or someone in their immediate family is serving or has ever served in the U.S. military

voted

factor variable for validated turnout in general elections - takes values of "Voted", "No Record of Voting", and "No Voter File"

Source

https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/II2DB6

Author

David Kane