Election results and candidate longevity

This data is from the paper "Longevity Returns to Political Office" by Barfort, Klemmensen & Larsen (2019). The purpose of this study was to explore how winning an election influences the lifespan of a candidate. The authors collected data on all U.S. gubernatorial elections from 1945 to 2012.

governors

Format

A tibble with 1,092 observations and 11 variables:

state

character variable indicating the state in which an election took place

year

integer variable indicating the year in which an election took place

first_name

character variable indicating the first name of a candidate

last_name

character variable indicating the last name of a candidate

party

character variable indicating a candidate's party

sex

character variable indicating a candidate's sex

died

date variable indicating the candidate's date of death

status

character variable indicating whether a candidate was the challenger or incumbent

win_margin

double variable indicating the percentage margin by which the election was won (positive values) or lost (negative values)

population

number of people living in the state at the time of the election

death_age

years a candidate lived

election_age

years a candidate lived before the election took place

lived_after

years a candidate lived after the election took place

Source

https://doi.org/10.7910/DVN/IBKYRX

Details

First, for a given election, only the two candidates who received the highest number of votes were included. Second, candidates with unknown dates of death were excluded, resulting in fewer observations for elections in recent years, since most recent candidates are still alive. Third, in a few instances, only the year of birth or death could be determined; in these cases, the date was taken to be July 1 of that year.

Table: Data summary

Name	governors

Number of rows	1092

Number of columns	14

_______________________

Column type frequency:

character	7

Date	1

numeric	6

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

state	0	1	4	14	0	50	0

first_name	0	1	2	11	0	299	0

last_name	0	1	3	11	0	615	0

party	0	1	8	11	0	3	0

sex	0	1	4	6	0	2	0

status	0	1	9	10	0	2	0

region	0	1	4	9	0	4	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

died	0	1	1946-12-21	2019-08-18	1996-07-20	683

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

year	0	1	1964.85	13.38	1945.00	1954.00	1.962e+03	1.974e+03	2011.00	▇▆▃▂▁

win_margin	0	1	3.22	26.07	-82.43	-9.87	9.700e-01	1.218e+01	100.00	▁▃▇▁▁

population	0	1	3783555.08	4231731.71	145000.00	900000.00	2.427e+06	4.536e+06	27102238.00	▇▁▁▁▁

election_age	0	1	51.72	8.71	31.35	45.34	5.136e+01	5.748e+01	83.87	▂▇▆▂▁

death_age	0	1	79.95	11.16	41.63	73.38	8.190e+01	8.815e+01	101.72	▁▂▅▇▃

lived_after	0	1	28.23	13.38	0.13	17.57	2.960e+01	3.867e+01	60.42	▃▆▇▆▂

Author

David Kane