American National Election Studies

This dataset is from the American National Election Studies (ANES), a survey that aims to provide insights into voter behavior. It has been conducted since 1948 before and after each presidential election, and combines questions about voters' political attitudes with extensive biographical information. Current as of 01/20/2021.

nes

Format

A tibble with 38,558 observations and 17 variables:

year

integer variable for year of study

state

character variable for the 2-letter abbreviation of respondent's survey state

gender

character variable with values "Male", "Female", and "Other"

income

factor variable for respondent's income percentile group. Has levels "0 - 16", "17 - 33", "34 - 67", "68 - 95", and "96 - 100"

age

factor variable for respondent's age group. Has levels "17 - 24", "25 - 34", "35 - 44", "45 - 54", "55 - 64", "65 - 74", and "75 +"

education

factor variable with a 7 levels measuring highest educational achievement

race

character variable for respondent's racial / ethnical identity

ideology

factor variable for respondent's party identification with 7 levels

voted

character variable indicating whether the respondent voted in the national elections

region

factor variable for the region corresponding to respondent's survey state

pres_appr

character variable for respondent's self-reported approval of the sitting president. Question was not asked before 1972

influence

character variable for respondent's opinion on the statement "People like me have any say in what the government does"

equality

character variable for respondent's opinion on the statement "Our society should do whatever is necessary to make sure that everyone has an equal opportunity to succeed". Question was not asked before 1984

religion

character variable for importance of religion in respondent's life. Question was not asked before 1980

better_alone

character variable for respondent's opinion on the statement "This country would be better off if we just stayed home and did not concern ourselves with problems in other parts of the world". Question was not asked before 1956

therm_black

integer variable for respondent's self-reported attitude toward black people. Takes "thermometer" values from 0 (very negative) to 100 (very positive). Question was not asked before 1964

therm_white

integer variable for respondent's self-reported attitude toward white people. Takes "thermometer" values from 0 (very negative) to 100 (very positive). Question was not asked before 1964

Source

https://electionstudies.org/data-center/anes-time-series-cumulative-data-file/

Details

Some of the questions aksed in the survey have changed slightly over time. The ANES codebook provides further information on this issue.

Author

David Kane