Data on Nobel Prize laureates

This data comes from the freely accessible API of the Nobel Foundation. The data includes information on all Nobel Prize laureates since 1901, as well as on all laureates in the Nobel Memorial Prize in Economic Sciences since 1968.

Table: Data summary

Name	nobel

Number of rows	961

Number of columns	16

_______________________

Column type frequency:

Date	2

character	10

factor	2

numeric	2

________________________

Group variables	None

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

born	33	0.97	1817-11-30	1997-07-12	1917-09-19	910

died	319	0.67	1903-11-01	2021-07-23	1984-10-17	631

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

first_name	0	1.00	2	59	0	734	0

last_name	30	0.97	2	26	0	876	0

born_country	28	0.97	3	52	0	124	0

born_city	30	0.97	3	36	0	630	0

died_country	332	0.65	3	35	0	52	0

died_city	338	0.65	4	30	0	312	0

motivation	0	1.00	22	341	0	671	0

aff_inst	0	1.00	4	137	0	343	0

aff_city	0	1.00	4	33	0	233	0

aff_country	28	0.97	3	52	0	124	0

Variable type: factor

skim_variable	n_missing	complete_rate	ordered	n_unique	top_counts

field	0	1	FALSE	6	Med: 222, Phy: 215, Che: 186, Pea: 135

gender	0	1	FALSE	3	Mal: 875, Fem: 58, Org: 28

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

year	0	1	1971.89	33.76	1901	1948	1977	2001	2020	▃▃▅▇▇

share	0	1	2.01	0.94	1	1	2	3	4	▇▇▁▅▂

nobel

Format

A tibble with 961 observations and 16 variables:

first_name

character variable for laureate's first name

last_name

character variable for laureate's last name

year

interger variable for year in which the price was awarded

field

factor variable for field in which the price was awarded

share

integer variable for number of laureates who shared the price

gender

factor variable for laureate's gender

born

date variable for laureate's birth date

died

date variable for laureate's death date

born_country

character variable for laureate's country of birth at the time

born_city

character variable for laureate's city of birth at the time

died_country

character variable for laureate's country of death at the time

died_city

character variable for laureate's city of death at the time

motivation

character variable for reason of award

aff_inst

character variable for laureate's academic affiliation

aff_city

character variable for city of institution of laureate's academic affiliation

aff_country

character variable for country of institution of laureate's academic affiliation

Source

https://www.nobelprize.org/about/developer-zone-2/

Author

David Kane