American National Election Studies

This dataset is from the American National Election Studies (ANES), a survey that aims to provide insights into voter behavior. It has been conducted since 1948 before and after each presidential election, and combines questions about voters' political attitudes with extensive biographical information. Current as of 01/20/2021.

nes

Format

A tibble with 38,558 observations and 17 variables:

year

integer variable for year of study

state

character variable for the 2-letter abbreviation of respondent's survey state

gender

character variable with values "Male", "Female", and "Other"

income

factor variable for respondent's income percentile group. Has levels "0 - 16", "17 - 33", "34 - 67", "68 - 95", and "96 - 100"

age

factor variable for respondent's age group. Has levels "17 - 24", "25 - 34", "35 - 44", "45 - 54", "55 - 64", "65 - 74", and "75 +"

education

factor variable with a 7 levels measuring highest educational achievement

race

character variable for respondent's racial / ethnical identity

ideology

factor variable for respondent's party identification with 7 levels

voted

character variable indicating whether the respondent voted in the national elections

region

factor variable for the region corresponding to respondent's survey state

pres_appr

character variable for respondent's self-reported approval of the sitting president. Question was not asked before 1972

influence

character variable for respondent's opinion on the statement "People like me have any say in what the government does"

equality

character variable for respondent's opinion on the statement "Our society should do whatever is necessary to make sure that everyone has an equal opportunity to succeed". Question was not asked before 1984

religion

character variable for importance of religion in respondent's life. Question was not asked before 1980

better_alone

character variable for respondent's opinion on the statement "This country would be better off if we just stayed home and did not concern ourselves with problems in other parts of the world". Question was not asked before 1956

therm_black

integer variable for respondent's self-reported attitude toward black people. Takes "thermometer" values from 0 (very negative) to 100 (very positive). Question was not asked before 1964

therm_white

integer variable for respondent's self-reported attitude toward white people. Takes "thermometer" values from 0 (very negative) to 100 (very positive). Question was not asked before 1964

Source

https://electionstudies.org/data-center/anes-time-series-cumulative-data-file/

Details

Some of the questions aksed in the survey have changed slightly over time. The ANES codebook provides further information on this issue.

Table: Data summary

Name	nes

Number of rows	38558

Number of columns	17

_______________________

Column type frequency:

character	9

factor	5

numeric	3

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

state	110	1.00	2	2	0	50	0

gender	141	1.00	4	6	0	3	0

race	287	0.99	5	15	0	6	0

voted	4078	0.89	2	3	0	2	0

pres_appr	9646	0.75	6	10	0	3	0

influence	6925	0.82	5	26	0	3	0

equality	16902	0.56	10	26	0	6	0

religion	14052	0.64	9	11	0	2	0

better_alone	7794	0.80	5	8	0	2	0

Variable type: factor

skim_variable	n_missing	complete_rate	ordered	n_unique	top_counts

income	2517	0.93	FALSE	5	34 : 11740, 68 : 9974, 0 -: 6300, 17 : 6213

age	450	0.99	FALSE	7	25 : 7669, 35 : 7342, 45 : 6545, 55 : 6021

education	397	0.99	FALSE	7	Hig: 9269, Som: 8540, Col: 5540, Som: 4688

ideology	619	0.98	FALSE	7	Str: 7784, Wea: 7330, Wea: 4894, Str: 4807

region	0	1.00	FALSE	4	Sou: 13680, Mid: 10072, Wes: 7428, Nor: 7378

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

year	0	1.00	1989.40	20.50	1952	1972	1992	2012	2016	▃▃▃▃▇

therm_black	9418	0.76	65.74	20.64	0	50	65	85	97	▁▁▇▇▇

therm_white	9354	0.76	73.66	18.60	0	60	75	85	97	▁▁▃▅▇

Author

David Kane