Data on Nobel Prize laureates

This data comes from the freely accessible API of the Nobel Foundation. The data includes information on all Nobel Prize laureates since 1901, as well as on all laureates in the Nobel Memorial Prize in Economic Sciences since 1968.

nobel

Format

A tibble with 961 observations and 15 variables:

first_name

character variable for laureate's first name

last_name

character variable for laureate's last name

year

interger variable for year in which the price was awarded

field

factor variable for field in which the price was awarded

share

integer variable for number of laureates who shared the price

gender

factor variable for laureate's gender

born

date variable for laureate's birth date

died

date variable for laureate's death date

born_country

character variable for laureate's country of birth at the time

born_city

character variable for laureate's city of birth at the time

died_country

character variable for laureate's country of death at the time

died_city

character variable for laureate's city of death at the time

motivation

character variable for reason of award

aff_inst

character variable for laureate's academic affiliation

aff_city

character variable for city of institution of laureate's academic affiliation

Source

https://www.nobelprize.org/about/developer-zone-2/

Author

David Kane