Michigan voting experiment with social shaming

This is a dataset corresponding to the paper "Social Pressure and Voter Turnout: Evidence from a Large-Scale Field Experiment" by Gerber, Green, and Larimer (2008). See also the 'Details' section below. The aim of the study was to find out whether and to what extent people are motivated to vote by social pressure. To answer this question, the authors conducted a field experiment prior to the August 2006 primary election in Michigan. A total of 180,000 households were randomly assigned to either a control group or one of four treatment groups.

shaming

Format

A tibble with 344,084 observations and 15 variables:

cluster

character variable with cluster designation, which ranges from 1 to 10,000. See the documentation for extensive discussion about the clustering procedure

primary_06

0/1 integer variable indicating whether the respondent voted in the 2006 primary election

treatment

factor variable indicating which of the 5 treatments were employed in 2006, but before the primary election that year: 'Control', 'Civic Duty', 'Hawthorne', 'Self', or 'Neighbors'

sex

character variable with values "Male" and "Female"

age

integer variable indicating the respondent's age in 2006

primary_00

character variable indicating whether the respondent voted in the 2000 primary election

general_00

character variable indicating whether the respondent voted in the 2000 general election

primary_02

character variable indicating whether the respondent voted in the 2002 primary election

general_02

character variable indicating whether the respondent voted in the 2002 general election

primary_04

character variable indicating whether the respondent voted in the 2004 primary election

general_04

character variable indicating whether the respondent voted in the 2004 general election

hh_size

integer variable indicating the respondent's household size

hh_primary_04

numeric variable indicating the percent of household members who voted in the 2004 primary election. This and hh_general_04 seem suspicious

hh_general_04

numeric variable indicating the percent of household members who voted in the 2004 general election. Most be greater than zero since, in order to be included in the sample, a respondent must have voted in that election

neighbors

integer variable indicating the number of neighbors. This is, presumably, most relevant if the respondent was in the "Neighbors" group. neighbors is, in that case, the number of names listed on the mailing. Large majority of values is 21, which is the maximum number of names which could be printed on the mailing

Source

https://doi.org/10.1017/S000305540808009X

Details

The control group consisted of approximately 100,000 households and was observed without further intervention.The treatment groups consisted of about 20,000 households each, and were sent one mailing 11 days prior to the primary election. The first treatment group, named “Civic Duty”, received a letter that only reminded them to "do their civic duty and vote". The second treatment group, named "Hawthorne", received the same message with an additional notice that they are being studied by researchers. The letter sent to the third group, named "Self", included the content in the Hawthorne letter, but added a notice that every household member would be notified of each others' voting behavior after the election (this information is public). The last group, "Neighbors", finally listed not only the household's voting records but also the voting records of those nearby. As in the "Self" group, everyone on the list would be notified of their voting behavior after the primary.

Table: Data summary

Name	shaming

Number of rows	344084

Number of columns	15

_______________________

Column type frequency:

character	8

factor	1

numeric	6

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

cluster	0	1	1	5	0	10000	0

sex	0	1	4	6	0	2	0

primary_00	0	1	2	3	0	2	0

general_00	0	1	2	3	0	2	0

primary_02	0	1	2	3	0	2	0

general_02	0	1	2	3	0	2	0

primary_04	0	1	2	3	0	2	0

general_04	0	1	3	3	0	1	0

Variable type: factor

skim_variable	n_missing	complete_rate	ordered	n_unique	top_counts

treatment	0	1	FALSE	5	Con: 191243, Civ: 38218, Sel: 38218, Haw: 38204

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

primary_06	0	1	0.32	0.46	0.00	0.00	0.00	1.00	1	▇▁▁▁▃

age	0	1	49.79	14.45	20.00	41.00	50.00	59.00	106	▃▇▅▁▁

hh_size	0	1	2.18	0.79	1.00	2.00	2.00	2.00	8	▇▂▁▁▁

hh_primary_04	0	1	0.31	0.18	0.00	0.19	0.29	0.43	1	▇▇▅▁▁

hh_general_04	0	1	0.92	0.07	0.38	0.90	0.95	1.00	1	▁▁▁▂▇

neighbors	0	1	20.93	0.51	1.00	21.00	21.00	21.00	21	▁▁▁▁▇

Author

David Kane